src/agent/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes --no-annotate --no-emit-project --frozen -o requirements.txt
agent-framework-azure-ai==1.0.0rc5
agent-framework-core==1.0.0rc5
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
azure-ai-agents==1.2.0b5
azure-ai-agentserver-agentframework==1.0.0b17
azure-ai-agentserver-core==1.0.0b17
azure-ai-inference==1.0.0b9
azure-ai-projects==2.0.1
azure-common==1.1.28
azure-core==1.38.2
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.15.0
azure-identity==1.26.0b2
azure-monitor-opentelemetry==1.8.4
azure-monitor-opentelemetry-exporter==1.0.0b48
azure-search-documents==11.7.0b2
azure-storage-blob==12.29.0b1
cachetools==7.0.1
certifi==2026.2.25
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==46.0.5
distro==1.9.0
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.1
isodate==0.7.2
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.26.0
msal==1.35.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.1
oauthlib==3.3.1
openai==2.24.0
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-instrumentation-asgi==0.60b0
opentelemetry-instrumentation-dbapi==0.60b0
opentelemetry-instrumentation-django==0.60b0
opentelemetry-instrumentation-fastapi==0.60b0
opentelemetry-instrumentation-flask==0.60b0
opentelemetry-instrumentation-psycopg2==0.60b0
opentelemetry-instrumentation-requests==0.60b0
opentelemetry-instrumentation-urllib==0.60b0
opentelemetry-instrumentation-urllib3==0.60b0
opentelemetry-instrumentation-wsgi==0.60b0
opentelemetry-proto==1.39.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b0
packaging==24.2
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
pydantic==2.13.0b2
pydantic-core==2.42.0
pydantic-settings==2.13.1
pyjwt==2.11.0
python-dotenv==1.2.1
python-multipart==0.0.22
pywin32==311 ; sys_platform == 'win32'
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
sniffio==1.3.1
sse-starlette==3.2.0
starlette==1.0.0rc1
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
